Create a new pool for PHP5-FPM instead of modifying the default.
This commit updates the PHP5-FPM and Nginx configuration functions to
create a new pool instead of modifying the default one. We can control
how this pool is configured, namely we want PHP5-FPM to listen on a Unix
socket which Nginx can connect to.

Tested on Lucid and Precise with PPAs 'nginx/stable' and 'ondrej/php5'.

See #3 for credits and discussion. Thanks @intel352.