Don't install Ruby gems that are already installed (no version check).
When running a provisioning script multiple times, it gets extremely
annoying to wait for Git to re-download the source of a Gem and then
build and install it.

When running `gem install ...` it could also get very slow if the
packages index is refreshed on each run.

To avoid both situations, check if a version of a Gem file is already
installed and only proceed if none exists. In the future an option to
specify which version of a Gem is needed should be made available.

The best approach would be to consider using Bundler for managing the
global gemset as well. We could then build Gemfiles{,.lock} on the fly
and offset all of the hard work.